MV3.0, Extended Field Coding and ODO format support
Set Reader/Writer version R/W_T4T_Vno to 3.0
Extend all offsets, lengths, data pointers to 32-bit variables
to address memories larger than 32kB
Support Extended Field Coding by default if MLc>255 and/or MLe>256, even for MV2.0
For MV3.0 detected Tags, support ODO and DDO (BLE-TLV support) coding with
new ReadBinary and UpdateBinary INS code, Extended Data Structure
For addressing offsets below 0x8000, use MV2.0 commands for both MV2.0 and MV3.0 Tags
For addressing offsets above 0x8000 (i.e. MV3.0 Tag detected), use new MV3.0 INS code commands

+++ SN100x/src/nfa/rw/nfa_rw_act.cc
@@ -802,7 +802,7 @@ static void nfa_rw_handle_t2t_evt(tRW_EVENT event, tRW_DATA* p_rw_data) {
       break;
 
     case RW_T2T_TLV_DETECT_EVT: /* Lock control/Mem/Prop tlv detection complete
-                                   */
+                                 */
       nfa_rw_handle_tlv_detect(p_rw_data);
       break;
 
@@ -1780,8 +1780,7 @@ static tNFC_STATUS nfa_rw_start_ndef_write(void) {
       status = RW_T3tUpdateNDef(nfa_rw_cb.ndef_wr_len, nfa_rw_cb.p_ndef_wr_buf);
     } else if (NFC_PROTOCOL_ISO_DEP == protocol) {
       /* ISODEP/4A,4B- NFC-A or NFC-B */
-      status = RW_T4tUpdateNDef((uint16_t)nfa_rw_cb.ndef_wr_len,
-                                nfa_rw_cb.p_ndef_wr_buf);
+      status = RW_T4tUpdateNDef(nfa_rw_cb.ndef_wr_len, nfa_rw_cb.p_ndef_wr_buf);
     } else if (NFC_PROTOCOL_T5T == protocol) {
       /* ISO 15693 */
       status = RW_I93UpdateNDef((uint16_t)nfa_rw_cb.ndef_wr_len,
@@ -1874,8 +1873,8 @@ static bool nfa_rw_write_ndef(tNFA_RW_MSG* p_data) {
   } else if (nfa_rw_cb.ndef_st == NFA_RW_NDEF_ST_FALSE) {
     if (nfa_rw_cb.protocol == NFC_PROTOCOL_T1T) {
       /* For Type 1 tag, NDEF can be written on Initialized tag
-      *  Perform ndef detection first to check if tag is in Initialized state to
-      * Write NDEF */
+       *  Perform ndef detection first to check if tag is in Initialized state
+       * to Write NDEF */
       write_status = nfa_rw_start_ndef_detection();
     } else {
       /* Tag is not NDEF */
